+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- `common_elements` to determine what elements two arrays share.
 - Functions to prepare groups for different calculations.
 - `DriverCube` and `psi4_cube` for cube property calculations like electrostatic potentials.
 - `parse_cube` parser for Gaussian cube files.